Adjusted tryGetTokenString to get successful build
